+++ package.json
@@ -5,11 +5,11 @@
   "main": "index.js",
   "scripts": {
     "chopsticks": "npx @acala-network/chopsticks@latest -c ./configs/coretime-rococo.yml",
-    "zombienet": "./scripts/run_zombienet.sh minimal",
     "zombienet:full": "./scripts/run_zombienet.sh full",
     "chopsticks-init": "npx ts-node src/chopsticks.init.ts",
     "zombienet-init": "npx ts-node src/zombienet.init.ts",
     "zombienet-init:full": "npx ts-node src/zombienet.init.ts --fullNetwork",
+    "zombienet-init:local": "npx ts-node src/zombienet.init.ts --local",
     "prettier": "prettier --write src/",
     "lint": "eslint src --ext .ts,.tsx",
     "lint:fix": "eslint src --ext .ts,.tsx --fix"
