--- src/chopsticks.init.ts
@@ -1,86 +1,0 @@
-import { ApiPromise, WsProvider, Keyring } from "@polkadot/api";
-import { purchaseRegion, log } from "./common";
-import { cryptoWaitReady } from "@polkadot/util-crypto";
-import * as consts from "./consts";
-import { Timeslice } from "coretime-utils";
-
-const keyring = new Keyring({ type: "sr25519" });
-
-async function init() {
-  const coretimeWsProvider = new WsProvider("ws://127.0.0.1:8000");
-  const rococoWsProvider = new WsProvider("wss://rococo-rpc.polkadot.io/");
-
-  const coretimeApi = await ApiPromise.create({ provider: coretimeWsProvider });
-  const rococoApi = await ApiPromise.create({ provider: rococoWsProvider });
-
-  await startBulkSale(rococoApi, coretimeApi);
-}
-
-init().then(() => process.exit(0));
-
-async function startBulkSale(rococoApi: ApiPromise, coretimeApi: ApiPromise) {
-  const latestRcBlock = (await rococoApi.rpc.chain.getHeader()).number.toNumber();
-  await setStatus(coretimeApi, latestRcBlock);
-  log("status set");
-  await initializeSale(coretimeApi, latestRcBlock);
-  log("sale initialized");
-
-  const alice = keyring.addFromUri("//Alice");
-  await cryptoWaitReady();
-
-  // Alice buys a region.
-  await purchaseRegion(coretimeApi, alice);
-  log("Region purchased");
-}
-
-async function setStatus(coretimeApi: ApiPromise, latestRcBlock: number): Promise<any> {
-  const commitTimeslice = getLatestTimesliceReadyToCommit(latestRcBlock);
-
-  const status = {
-    core_count: consts.CORE_COUNT,
-    private_pool_size: 0,
-    system_pool_size: 0,
-    last_committed_timeslice: currentTimeslice(commitTimeslice) - 1,
-    last_timeslice: currentTimeslice(latestRcBlock),
-  };
-
-  await coretimeApi.rpc("dev_setStorage", {
-    broker: {
-      status,
-    },
-  });
-  return await coretimeApi.rpc("dev_newBlock");
-}
-
-async function initializeSale(coretimeApi: ApiPromise, latestRcBlock: number): Promise<any> {
-  const now = (await coretimeApi.rpc.chain.getHeader()).number.toNumber();
-  const commitTimeslice = getLatestTimesliceReadyToCommit(latestRcBlock);
-
-  const saleInfo = {
-    sale_start: now,
-    leadin_length: consts.CONFIG.leadin_length,
-    price: consts.INITIAL_PRICE,
-    sellout_price: null,
-    region_begin: commitTimeslice,
-    region_end: commitTimeslice + consts.CONFIG.region_length,
-    first_core: 0,
-    ideal_cores_sold: consts.IDEAL_CORES_SOLD,
-    cores_offered: consts.CORE_COUNT,
-    cores_sold: 0,
-  };
-
-  return await coretimeApi.rpc("dev_setStorage", {
-    broker: {
-      saleInfo,
-    },
-  });
-}
-
-function currentTimeslice(latestRcBlock: number) {
-  return Math.floor(latestRcBlock / consts.TIMESLICE_PERIOD);
-}
-
-function getLatestTimesliceReadyToCommit(latestRcBlock: number): Timeslice {
-  const advanced = latestRcBlock + consts.CONFIG.advance_notice;
-  return Math.floor(advanced / consts.TIMESLICE_PERIOD);
-}
