+++ src/zombienet.init.ts
@@ -1,390 +1,47 @@
-import { ApiPromise, WsProvider, Keyring } from "@polkadot/api";
-import { u8aToHex } from "@polkadot/util";
 import { cryptoWaitReady } from "@polkadot/util-crypto";
-import { purchaseRegion, log, normalizePath, transferRegion } from "./common";
-import { Abi, ContractPromise } from "@polkadot/api-contract";
 import { program } from "commander";
-import fs from "fs";
-import * as consts from "./consts";
-import { CoreMask, Id, Region } from "coretime-utils";
 import process from "process";
-import type { WeightV2 } from "@polkadot/types/interfaces";
-import { BN, bnToBn } from "@polkadot/util";
+import { relayInit } from "./relay/index.init";
+import { coretimeInit } from "./coretime/index.init";
+import { contractsInit } from "./contracts/index.init";
+import { normalizePath } from "./utils";
 
 program
-  .option("--fullNetwork")
-  .option("--contracts <string>")
+  .option("--dev")
+  .option("--relayInit")
+  .option("--coretimeInit")
+  .option("--contractsInit")
+  .option("--contractsPath <string>")
   .option("--contractsAccount <string>")
   .option("--coretimeAccount <string>");
 
 program.parse(process.argv);
 
-const REGION_COLLECTION_ID = 42;
-
-const CORETIME_CHAIN_PARA_ID = 1005;
-const CONTRACTS_CHAIN_PARA_ID = 2000;
+const CORETIME_PARA_ID = 1005;
+const CONTRACTS_PARA_ID = 2000;
 
 const ROCOCO_ENDPOINT = "ws://127.0.0.1:9900";
 const CORETIME_ENDPOINT = "ws://127.0.0.1:9910";
-const CONTRACTS_ENDPOINT = "ws://127.0.0.1:9920";
-
-const keyring = new Keyring({ type: "sr25519" });
+const CONTRACTS_ENDPOINT = program.opts().dev ? "ws://127.0.0.1:9944" : "ws://127.0.0.1:9920";
 
 async function init() {
-  const rococoWsProvider = new WsProvider(ROCOCO_ENDPOINT);
-  const coretimeWsProvider = new WsProvider(CORETIME_ENDPOINT);
-
-  const coretimeApi = await ApiPromise.create({ provider: coretimeWsProvider });
-  const rococoApi = await ApiPromise.create({ provider: rococoWsProvider });
-
   await cryptoWaitReady();
-  const alice = keyring.addFromUri("//Alice");
 
-  await configureBroker(coretimeApi);
-  await startSales(coretimeApi);
-
-  await setBalance(coretimeApi, alice.address, 1000 * consts.UNIT);
-
-  const regionId = await purchaseRegion(coretimeApi, alice);
-
-  const coretimeAccount = program.opts().coretimeAccount;
-  if (coretimeAccount) {
-    await transferRegion(coretimeApi, alice, coretimeAccount, regionId);
+  if (program.opts().relayInit) {
+    await relayInit(ROCOCO_ENDPOINT, CORETIME_PARA_ID, CONTRACTS_PARA_ID);
   }
 
-  if (program.opts().fullNetwork) {
-    const account = program.opts().contractsAccount;
+  if (program.opts().coretimeInit) {
+    await coretimeInit(CORETIME_ENDPOINT, program.opts().coretimeAccount || "");
+  }
 
-    await openHrmpChannel(rococoApi, CORETIME_CHAIN_PARA_ID, CONTRACTS_CHAIN_PARA_ID);
-
-    const contractsProvider = new WsProvider(CONTRACTS_ENDPOINT);
-    const contractsApi = await ApiPromise.create({ provider: contractsProvider, types: { Id } });
-
-    const xcRegionsAddress = await deployXcRegions(contractsApi);
-    log(`XcRegions address: ${xcRegionsAddress}`);
-
-    const marketAddress = await deployMarket(contractsApi, xcRegionsAddress);
-    log(`Market address: ${marketAddress}`);
-
-    await createRegionCollection(contractsApi);
-
-    const mockRegion = new Region(
-      { begin: 30, core: 0, mask: CoreMask.fromChunk(0, 40) },
-      { end: 60, owner: alice.address, paid: null }
-    );
-
-    await mintRegion(contractsApi, mockRegion);
-    await approveTransfer(contractsApi, mockRegion, xcRegionsAddress);
-    await initXcRegion(contractsApi, xcRegionsAddress, mockRegion);
-    if (account) {
-      await transferWrappedRegion(contractsApi, xcRegionsAddress, mockRegion, account);
+  if (program.opts().contractsInit) {
+    if(!program.opts().contractsPath) {
+      throw new Error("--contractsPath must be specified");
     }
+    const contractsPath = normalizePath(program.opts().contractsPath);
+    await contractsInit(CONTRACTS_ENDPOINT, program.opts().contractsAccount || "", contractsPath);
   }
 }
 
 init().then(() => process.exit(0));
-
-async function configureBroker(coretimeApi: ApiPromise): Promise<void> {
-  log(`Setting the initial configuration for the broker pallet`);
-
-  const configCall = coretimeApi.tx.broker.configure(consts.CONFIG);
-  return force(coretimeApi, configCall);
-}
-
-async function startSales(coretimeApi: ApiPromise): Promise<void> {
-  log(`Starting the bulk sale`);
-
-  const startSaleCall = coretimeApi.tx.broker.startSales(consts.INITIAL_PRICE, consts.CORE_COUNT);
-  return force(coretimeApi, startSaleCall);
-}
-
-async function setBalance(coretimeApi: ApiPromise, who: string, balance: number) {
-  log(`Setting balance of ${who} to ${balance}`);
-
-  const setBalanceCall = coretimeApi.tx.balances.forceSetBalance(who, balance);
-  return force(coretimeApi, setBalanceCall);
-}
-
-async function openHrmpChannel(rococoApi: ApiPromise, sender: number, recipient: number): Promise<void> {
-  log(`Openeing HRMP channel between ${sender} - ${recipient}`);
-
-  const newHrmpChannel = [
-    sender,
-    recipient,
-    8, // Max capacity
-    512, // Max message size
-  ];
-
-  const alice = keyring.addFromUri("//Alice");
-
-  const openHrmp = rococoApi.tx.hrmp.forceOpenHrmpChannel(...newHrmpChannel);
-  const sudoCall = rococoApi.tx.sudo.sudo(openHrmp);
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await sudoCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function deployXcRegions(contractsApi: ApiPromise): Promise<string> {
-  log(`Uploading xcRegions contract code`);
-  const contractsPath = normalizePath(program.opts().contracts);
-
-  const wasm = getXcRegionsWasm(contractsPath);
-  const metadata = getXcRegionsMetadata(contractsApi, contractsPath);
-
-  return Promise.resolve(instantiateWithCode(contractsApi, wasm, metadata, []));
-}
-
-async function deployMarket(contractsApi: ApiPromise, xcRegionsContract: string): Promise<string> {
-  log(`Uploading market contract code`);
-  const contractsPath = normalizePath(program.opts().contracts);
-
-  const wasm = getMarketWasm(contractsPath);
-  const metadata = getMarketMetadata(contractsApi, contractsPath);
-  const listingDeposit = 0;
-  const timeslicePeriod = 80;
-
-  return Promise.resolve(
-    instantiateWithCode(contractsApi, wasm, metadata, [xcRegionsContract, listingDeposit, timeslicePeriod])
-  );
-}
-
-async function instantiateWithCode(
-  contractsApi: ApiPromise,
-  wasm: Buffer,
-  metadata: Abi,
-  params: any[]
-): Promise<string> {
-  const alice = keyring.addFromUri("//Alice");
-
-  const value = 0;
-  const storageDepositLimit = null;
-
-  const instantiate = contractsApi.tx.contracts.instantiateWithCode(
-    value,
-    getMaxGasLimit(),
-    storageDepositLimit,
-    u8aToHex(wasm),
-    metadata.findConstructor(0).toU8a(params),
-    null
-  );
-
-  const callTx = async (resolve: (address: string) => void) => {
-    const unsub = await instantiate.signAndSend(alice, async (result: any) => {
-      if (result.status.isInBlock) {
-        const address = await getContractAddress(contractsApi);
-        unsub();
-        resolve(address);
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-// Create a mock collection that will represent regions.
-async function createRegionCollection(contractsApi: ApiPromise): Promise<void> {
-  log(`Creating the region collection`);
-
-  const alice = keyring.addFromUri("//Alice");
-  const createCollectionCall = contractsApi.tx.uniques.create(REGION_COLLECTION_ID, alice.address);
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await createCollectionCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function mintRegion(contractsApi: ApiPromise, region: Region): Promise<void> {
-  log(`Minting a region`);
-
-  const alice = keyring.addFromUri("//Alice");
-  const rawRegionId = region.getEncodedRegionId(contractsApi);
-  const mintCall = contractsApi.tx.uniques.mint(REGION_COLLECTION_ID, rawRegionId, alice.address);
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await mintCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function approveTransfer(contractsApi: ApiPromise, region: Region, delegate: string): Promise<void> {
-  log(`Approving region to ${delegate}`);
-
-  const alice = keyring.addFromUri("//Alice");
-  const rawRegionId = region.getEncodedRegionId(contractsApi);
-  const approveCall = contractsApi.tx.uniques.approveTransfer(REGION_COLLECTION_ID, rawRegionId, delegate);
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await approveCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function initXcRegion(contractsApi: ApiPromise, contractAddress: string, region: Region): Promise<void> {
-  log(`Initializing the metadata for a xc-region`);
-
-  const contractsPath = normalizePath(program.opts().contracts);
-
-  const metadata = getXcRegionsMetadata(contractsApi, contractsPath);
-  const xcRegionsContract = new ContractPromise(contractsApi, metadata, contractAddress);
-
-  const rawRegionId = region.getEncodedRegionId(contractsApi);
-  const id = contractsApi.createType("Id", { U128: rawRegionId });
-
-  const alice = keyring.addFromUri("//Alice");
-
-  const callArguments = [
-    id,
-    // All the region metadata combined:
-    {
-      begin: region.getBegin(),
-      end: region.getEnd(),
-      core: region.getCore(),
-      mask: region.getMask().getMask(),
-    },
-  ];
-
-  const initCall = xcRegionsContract.tx["regionMetadata::init"](
-    { gasLimit: getGasLimit(contractsApi, "8000000000", "250000"), storageDepositLimit: null },
-    ...callArguments
-  );
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await initCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function transferWrappedRegion(
-  contractsApi: ApiPromise,
-  contractAddress: string,
-  region: Region,
-  receiver: string
-): Promise<void> {
-  log(`Transferring wrapped region to ${receiver}`);
-
-  const contractsPath = normalizePath(program.opts().contracts);
-
-  const metadata = getXcRegionsMetadata(contractsApi, contractsPath);
-  const xcRegionsContract = new ContractPromise(contractsApi, metadata, contractAddress);
-
-  const rawRegionId = region.getEncodedRegionId(contractsApi);
-
-  const alice = keyring.addFromUri("//Alice");
-
-  const id = contractsApi.createType("Id", { U128: rawRegionId });
-  const callArguments = [receiver, id, []];
-
-  const transferCall = xcRegionsContract.tx["psp34::transfer"](
-    { gasLimit: getGasLimit(contractsApi, "8000000000", "250000"), storageDepositLimit: null },
-    ...callArguments
-  );
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await transferCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function force(api: ApiPromise, call: any): Promise<void> {
-  const sudoCall = api.tx.sudo.sudo(call);
-
-  const alice = keyring.addFromUri("//Alice");
-
-  const callTx = async (resolve: () => void) => {
-    const unsub = await sudoCall.signAndSend(alice, (result: any) => {
-      if (result.status.isInBlock) {
-        unsub();
-        resolve();
-      }
-    });
-  };
-
-  return new Promise(callTx);
-}
-
-async function getContractAddress(contractsApi: ApiPromise): Promise<string> {
-  log("Getting contract address");
-  const events: any = await contractsApi.query.system.events();
-
-  for (const record of events) {
-    const { event } = record;
-    if (event.section === "contracts" && event.method === "Instantiated") {
-      return event.data[1].toString();
-    }
-  }
-
-  return "";
-}
-
-const getMaxGasLimit = () => {
-  return {
-    refTime: 5000000000,
-    proofSize: 900000,
-  };
-};
-
-export const getGasLimit = (api: ApiPromise, _refTime: string | BN, _proofSize: string | BN): WeightV2 => {
-  const refTime = bnToBn(_refTime);
-  const proofSize = bnToBn(_proofSize);
-
-  return api.registry.createType("WeightV2", {
-    refTime,
-    proofSize,
-  }) as WeightV2;
-};
-
-const getXcRegionsMetadata = (contractsApi: ApiPromise, contractsPath: string) =>
-  new Abi(
-    fs.readFileSync(`${contractsPath}/xc_regions/xc_regions.json`, "utf-8"),
-    contractsApi.registry.getChainProperties()
-  );
-
-const getMarketMetadata = (contractsApi: ApiPromise, contractsPath: string) =>
-  new Abi(
-    fs.readFileSync(`${contractsPath}/coretime_market/coretime_market.json`, "utf-8"),
-    contractsApi.registry.getChainProperties()
-  );
-
-const getXcRegionsWasm = (contractsPath: string) => fs.readFileSync(`${contractsPath}/xc_regions/xc_regions.wasm`);
-
-const getMarketWasm = (contractsPath: string) =>
-  fs.readFileSync(`${contractsPath}/coretime_market/coretime_market.wasm`);
